+++ backend/server/ml_estimators/benchmark_estimator/__init__.py
@@ -1,0 +1,1 @@
+from .benchmark_estimator import BenchmarkEstimator
