--- backend/server/ml_models/ensemble_model/__init__.py
@@ -1,1 +1,0 @@
-from .ensemble_model import EnsembleModel
