+++ backend/server/models.py
@@ -5,7 +5,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
 from sklearn.externals import joblib
 
-from server.ml_models import BaseMLModel
+from server.ml_estimators import BaseMLEstimator
 from server.data_config import TEAM_NAMES
 
 # Rough estimate, but exactitude isn't necessary here
@@ -97,7 +97,7 @@ class MLModel(models.Model):
         max_length=500, null=True, blank=True, validators=[validate_module_path]
     )
 
-    def load_estimator(self) -> BaseMLModel:
+    def load_estimator(self) -> BaseMLEstimator:
         return joblib.load(self.filepath)
